requirements.txt
streamlit==1.32.0
torch==2.0.1
torchvision==0.15.2
timm==0.9.2
opencv-python-headless==4.8.0.74
pillow==9.5.0
numpy==1.24.3
matplotlib==3.7.1